Create diceRollWithRolzAPI.gs
This is Chad,

This is my more advance dice roller, but I kind of cheated using the rolz.org API. I found it easier than trying to figure out all the coding for all the different kids of rolls I wanted. For this you just type "!dice [type of roll]" example "!dice 2d10 +5". the bot will return with something like "21 = [ 10 +6 ] +5" This will be great for roleplaying for those that like it. I sure there is a better way to code this but it all made sense to me and it works. Any and all critiques are welcome. Hope you guys like it.